refactor: metric explorer (#40)
* refactor: restructure metrics explorer statistics components

* moves pagination to a separate component

* moves get_histogram to a lib

* avoid using string enum value for embedding type

* split and decouple similarities and embeddings from explorer
